requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Werkzeug==3.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
PyMySQL==1.1.0
pandas==2.2.1
openpyxl==3.1.2
reportlab==4.1.0
Pillow==10.2.0
img2pdf==0.5.1 